requirements.txt
azure-storage-queue 12.1.4
azure-storage-file-share 12.3.0
azure-storage-blob 12.6.0
azure-keyvault 4.1.0